+++ pipelines/matrix/src/matrix/datasets/drp_model.py
@@ -11,7 +11,7 @@ from typing import Any, Dict, List, Union, Tuple
 from refit.v1.core.make_list_regexable import make_list_regexable
 
 from matrix.datasets.graph import KnowledgeGraph
-from matrix.pipelines.modelling.nodes import _add_embeddings
+from matrix.pipelines.modelling.utils import _add_embeddings
 
 
 class DRPmodel(abc.ABC):
@@ -95,7 +95,7 @@ class DRPmodel3classScikit(DRPmodel3class):
         pairs = pairs.copy()
 
         # Add embeddings
-        pairs = _add_embeddings(pairs, graph)
+        pairs = _add_embeddings(pairs, self.graph)
 
         # Iterate over transformers
         for _, transform in self.transformers.items():
